+++ src/rics/utility/__init__.py
@@ -1,6 +1,13 @@
 """Utility modules."""
 
-from rics.utility import collections, logs, misc, perf, plotting, strings
+from rics.utility import collections, logs, misc, perf, strings
 from rics.utility._just_the_way_i_like_it import configure_stuff
 
-__all__ = ["configure_stuff", "collections", "logs", "misc", "perf", "strings", "plotting"]
+__all__ = ["configure_stuff", "collections", "logs", "misc", "perf", "strings"]
+
+try:
+    from rics.utility import plotting
+
+    __all__ += ["plotting"]
+except ModuleNotFoundError:
+    pass